requirements.txt
pandas
numpy
awscli
boto3
scikit-learn
scipy
matplotlib
seaborn
pymongo
certifi
python-dotenv
